+++ backend/api_gateway/internal/usecases/timeSlots/timeSlots_dto.go
@@ -10,8 +10,8 @@ import (
 type CreateTimeSlotReq struct {
 	Title      string                   `json:"title" validate:"required,max=255"`
 	Type       string                   `json:"type" validate:"required,oneof=single recurring"`
-	LocationID int                      `json:"locationId" validate:"required"`
-	Capacity   int                      `json:"capacity" validate:"required,min=1"`
+	LocationID int64                    `json:"locationId" validate:"required"`
+	Capacity   int32                    `json:"capacity" validate:"required,min=1"`
 	StartDate  time.Time                `json:"startDate" validate:"required"`
 	EndDate    time.Time                `json:"endDate" validate:"required,gtfield=StartDate"`
 	Services   []models.TimeSlotService `json:"services" validate:"required,dive,min=1"`
